+++ NLayerApp/NLayerApp.BusinessLogicLayer/Models/CommercialViewModel.cs
@@ -16,7 +16,7 @@ namespace NLayerApp.BusinessLogicLayer.Models
         public string OperationType { get; set; }
         public string NameInfo { get; set; }
         public string StateCommercial { get; set; }
-        public float TotalArea { get; set; }
+        public double TotalArea { get; set; }
         public int DollarPrice{ get; set; }
         public int GrnPrice{ get; set; }
     }
