+++ NLayerApp/NLayerApp.BusinessLogicLayer/Models/HouseDetailsInfoViewModel.cs
@@ -27,7 +27,7 @@ namespace NLayerApp.BusinessLogicLayer.Models
         public int FloorHouse { get; set; } //
         public string FloorMaterial { get; set; }
         public string WallMaterial { get; set; }
-        public float TotalAreaInfo { get; set; }
+        public double TotalAreaInfo { get; set; }
         public float LivingAreaHouse { get; set; } //
         public float KitchenAreaHouse { get; set; } //
         public float LandAreaHouse { get; set; } //
